feat: Add cursor pointer to Button component
This commit adds the `cursor: pointer` style to the Button component in the frontend. The change is made in the `Button.scss` file, adding the `cursor: pointer` property to the `.app__button` class. This modification allows the button to display a pointer cursor when hovered over, providing visual feedback to users that the button is clickable.